requirements.txt
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
colorama>=0.4.6
tqdm>=4.65.0
feedparser>=6.0.0
